Almost work with separate prepare task for Vault generation (todo fix vaultInputDir of aemCompose)
